+++ containers/daap-python-base/CHANGELOG.md
@@ -12,9 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- data_product_metadata python module. This contains code that will be used, intially
-by an api endpoint to create data product metadata, but in future, expanded to be used
-in other endpoints too, such as an update_metadata endpoint.
+- data_product_metadata python module. This contains code that will be used,
+intially by an API endpoint to create data product metadata, but in future,
+expanded to be usedin other endpoints too, such as an update_metadata endpoint.
 
 ## [0.3.0] - 2023-08-11